AG-17112 move toolbarMenuBuilder into MenuCoreModule
Relocate ToolbarMenuBuilder from the toolbar package to the menu
package and register it on MenuCoreModule's beans instead of
ToolbarModule's. ToolbarModule no longer statically imports
MenuList/MenuItemMapper, so consumers that do not register
ContextMenuModule or ColumnMenuModule can tree-shake the menu
widgets out of their bundle. agMenuToolbarItem remains registered
on ToolbarModule and looks up toolbarMenuBuilder at runtime,
falling back on the existing ContextMenu/ColumnMenu assertion when
neither module is registered.
