+++ packages/ag-grid-enterprise/src/toolbar/toolbarModule.ts
@@ -9,7 +9,6 @@ import { MenuToolbarItem } from './providedItems/menuToolbarItem';
 import { PivotPanelToolbarItem } from './providedItems/pivotPanelToolbarItem';
 import { QuickFilterToolbarItem } from './providedItems/quickFilterToolbarItem';
 import { RowGroupPanelToolbarItem } from './providedItems/rowGroupPanelToolbarItem';
-import { ToolbarMenuBuilder } from './providedItems/toolbarMenuBuilder';
 
 /**
  * @feature Accessories -> Toolbar
@@ -18,7 +17,6 @@ import { ToolbarMenuBuilder } from './providedItems/toolbarMenuBuilder';
 export const ToolbarModule: _ModuleWithoutApi = {
     moduleName: 'Toolbar',
     version: VERSION,
-    beans: [ToolbarMenuBuilder],
     userComponents: {
         agButtonToolbarItem: ButtonToolbarItem,
         agFindToolbarItem: FindToolbarItem,
